+++ simplerogue/src/main/java/game/Dungeon.java
@@ -6,7 +6,6 @@ import java.util.function.Supplier;
 
 import game.display.Display;
 import game.display.FloorMenu;
-import game.floorgeneration.DefaultFloorGenerator;
 import game.floorgeneration.Pool;
 import game.floorgeneration.Pools;
 import game.gameobjects.Floor;
@@ -17,7 +16,7 @@ import game.gameobjects.entities.PlayerEntity;
 import game.gameobjects.entities.Wall;
 import game.gameobjects.items.Item;
 import game.gameobjects.terrains.Staircase;
-import game.floorgeneration.DefaultFloorGeneratorImproved;
+import game.floorgeneration.DefaultFloorGenerator;
 
 public class Dungeon {
 
@@ -73,7 +72,7 @@ public class Dungeon {
         currentDepth = 1;
         sX = sizeX;
         sY = sizeY;
-        currentFloor = new Floor(sX, sY, new DefaultFloorGeneratorImproved(currentDepth));
+        currentFloor = new Floor(sX, sY, new DefaultFloorGenerator(currentDepth));
     }
     
     public static void update(){
