+++ simplerogue/src/main/java/game/PathConditions.java
@@ -20,11 +20,11 @@ public class PathConditions {
     private List<Function<Space,Double>> deterrentConditions = new ArrayList<Function<Space,Double>>();
 
     private boolean diagonal = true;
-  
+
     private BiFunction<Node,Node,Double> hFunction = (from, to) -> {
         return (double)Math.abs(to.x - from.x) + Math.abs(to.y - from.y);
     };
-  
+
     public PathConditions() {
         super();
     }
@@ -88,7 +88,7 @@ public class PathConditions {
         this.diagonal = diagonal;
         return this;
     }
-  
+
     public BiFunction<Node,Node,Double> getHFunction(){
         return this.hFunction;
     }
@@ -97,5 +97,5 @@ public class PathConditions {
         this.hFunction = hFunction;
         return this;
     }
-  
+
 }
