+++ simplerogue/src/main/java/game/display/Display.java
@@ -40,7 +40,7 @@ public class Display {
 
     private static Menu currentMenu;
     private static FloorMenu rootMenu;
-    // private static Mode mode = Mode.GRAPHICAL;
+    //private static Mode mode = Mode.GRAPHICAL;
     private static Mode mode = Mode.ASCII;
 
     private static KeyMap keyMap = new KeyMap();
